Fold the embed helpers into the files that already own them.
The extra modules and characterization tests did not change the cookie-plant path.
