--- src/shared/lib/cloud-dashboard-session-schema.ts
@@ -1,8 +1,0 @@
-import { z } from 'zod'
-
-export const CloudDashboardSessionSchema = z.object({
-  useCloudOrigin: z.boolean(),
-  origin: z.string().nullable(),
-})
-
-export type CloudDashboardSession = z.infer<typeof CloudDashboardSessionSchema>
